requirements.txt
cbor==1.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
oisp==0.2.1
paho-mqtt==1.6.1
Pygments==2.15.1
requests==2.31.0
termcolor==2.3.0
urllib3==2.0.3
